backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1
redis==5.2.1
python-dotenv==1.0.1
httpx==0.28.1
psycopg[binary,pool]==3.2.3
